fix(ui): extract update panel into shared component, fix empty modal on startup
UpdateAvailableModal was showing blank content on app start because its
standalone modal container had no defined height, causing flex:1 on the
content view to collapse to zero.

Extracted all update states (checking, available, up_to_date, error) into
UpdateContentPanel — a single source of truth that mirrors the Settings
subpanel exactly. UpdateAvailableModal is now a thin shell. Removed ~150
lines of duplicated code (stripMarkdown, formatApkDate, styles) from both
SettingsModal and UpdateAvailableModal.

🧀
