+++ app/modals/UpdateAvailableModal.js
@@ -1,43 +1,20 @@
-import React, { useRef, useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import { View, StyleSheet, TouchableOpacity, Animated, Easing, ScrollView } from 'react-native';
-import { Portal, Modal, Text, Divider, TouchableRipple } from 'react-native-paper';
+import { View, StyleSheet, TouchableOpacity } from 'react-native';
+import { Portal, Modal, Text, Divider } from 'react-native-paper';
 import { Ionicons } from '@expo/vector-icons';
 import { useThemeColors } from '../contexts/ThemeColorsContext';
 import { useLocalization } from '../contexts/LocalizationContext';
 import { HORIZONTAL_PADDING, SPACING, BORDER_RADIUS } from '../styles/layout';
-
-const stripMarkdown = (md) => md
-  .replace(/\r\n/g, '\n')
-  .replace(/#{1,6}\s+/g, '')
-  .replace(/\*\*(.+?)\*\*/g, '$1')
-  .replace(/\*(.+?)\*/g, '$1')
-  .replace(/\[([^\]]+)\]\([^)]+\)/g, '$1')
-  .replace(/`([^`]+)`/g, '$1')
-  .replace(/^\s*[-*+]\s+/gm, '• ')
-  .replace(/\n{3,}/g, '\n\n')
-  .trim();
+import UpdateContentPanel from '../components/UpdateContentPanel';
 
 export default function UpdateAvailableModal({ visible, onDismiss, onUpdate, updateData }) {
   const { colors } = useThemeColors();
   const { t } = useLocalization();
-  const contentAnim = useRef(new Animated.Value(0)).current;
-
-  useEffect(() => {
-    if (visible) {
-      contentAnim.setValue(0);
-      Animated.timing(contentAnim, {
-        toValue: 1,
-        duration: 280,
-        easing: Easing.out(Easing.quad),
-        useNativeDriver: true,
-      }).start();
-    }
-  }, [visible, contentAnim]);
 
   if (!updateData) return null;
 
-  const { latestVersion, currentVersion, downloadUrl, releaseNotes } = updateData;
+  const updateResult = { type: 'available', ...updateData };
 
   return (
     <Portal>
@@ -52,66 +29,13 @@ export default function UpdateAvailableModal({ visible, onDismiss, onUpdate, upd
             </Text>
             <View style={styles.backButton} />
           </View>
-
           <Divider />
-
-          <Animated.View style={[styles.resultContainer, { opacity: contentAnim }]}>
-            <View style={styles.updateAvailableHeader}>
-              <Ionicons name="download-outline" size={36} color={colors.primary} />
-              <View style={styles.updateVersionInfo}>
-                <Text style={[styles.updateNewVersion, { color: colors.text }]}>
-                  v{latestVersion}
-                </Text>
-                <Text style={[styles.updateCurrentVersion, { color: colors.mutedText }]}>
-                  {(t('update_from_version') || 'installed: v{currentVersion}')
-                    .replace('{currentVersion}', currentVersion)}
-                </Text>
-              </View>
-            </View>
-
-            {releaseNotes ? (
-              <>
-                <Divider style={styles.updateDivider} />
-                <Text style={[styles.changelogTitle, { color: colors.mutedText }]}>
-                  {t('whats_new') || "What's new"}
-                </Text>
-                <ScrollView style={styles.changelogScroll} showsVerticalScrollIndicator={false}>
-                  {releaseNotes.map(({ version, notes }) => (
-                    <View key={version} style={styles.changelogSection}>
-                      {releaseNotes.length > 1 && (
-                        <Text style={[styles.changelogVersion, { color: colors.mutedText }]}>
-                          v{version}
-                        </Text>
-                      )}
-                      <Text style={[styles.changelogText, { color: colors.text }]}>
-                        {stripMarkdown(notes)}
-                      </Text>
-                    </View>
-                  ))}
-                </ScrollView>
-              </>
-            ) : (
-              <Text style={[styles.updateVersionText, { color: colors.mutedText }]}>
-                {t('update_install_hint') || 'If installation is blocked, allow "Install unknown apps" for your browser or file manager in Android settings.'}
-              </Text>
-            )}
-
-            <View style={styles.updateActions}>
-              {releaseNotes && (
-                <Text style={[styles.updateHintText, { color: colors.mutedText }]}>
-                  {t('update_install_hint') || 'If installation is blocked, allow "Install unknown apps" for your browser or file manager in Android settings.'}
-                </Text>
-              )}
-              <TouchableRipple
-                onPress={() => onUpdate(downloadUrl)}
-                style={[styles.updateButton, { backgroundColor: colors.primary }]}
-              >
-                <Text style={styles.updateButtonText}>
-                  {t('update_now') || 'Update now'}
-                </Text>
-              </TouchableRipple>
-            </View>
-          </Animated.View>
+          <UpdateContentPanel
+            isChecking={false}
+            updateResult={updateResult}
+            downloadedApks={[]}
+            onUpdate={onUpdate}
+          />
         </View>
       </Modal>
     </Portal>
@@ -140,31 +64,6 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     width: 40,
   },
-  changelogScroll: {
-    flex: 1,
-    marginTop: SPACING.xs,
-  },
-  changelogSection: {
-    marginBottom: SPACING.md,
-  },
-  changelogText: {
-    fontSize: 13,
-    lineHeight: 20,
-  },
-  changelogTitle: {
-    fontSize: 11,
-    fontWeight: '600',
-    letterSpacing: 0.8,
-    marginTop: SPACING.md,
-    textTransform: 'uppercase',
-  },
-  changelogVersion: {
-    fontSize: 11,
-    fontWeight: '600',
-    letterSpacing: 0.5,
-    marginBottom: SPACING.xs,
-    textTransform: 'uppercase',
-  },
   header: {
     alignItems: 'center',
     flexDirection: 'row',
@@ -180,56 +79,5 @@ const styles = StyleSheet.create({
     margin: SPACING.md,
     maxHeight: '95%',
     overflow: 'hidden',
-  },
-  resultContainer: {
-    flex: 1,
-    paddingHorizontal: HORIZONTAL_PADDING,
-    paddingVertical: SPACING.lg,
-  },
-  updateActions: {
-    paddingTop: SPACING.md,
-  },
-  updateAvailableHeader: {
-    alignItems: 'center',
-    flexDirection: 'row',
-    gap: SPACING.md,
-    paddingBottom: SPACING.sm,
-  },
-  updateButton: {
-    borderRadius: BORDER_RADIUS.md,
-    marginTop: SPACING.xl,
-    paddingHorizontal: SPACING.xl,
-    paddingVertical: SPACING.md,
-  },
-  updateButtonText: {
-    color: '#fff',
-    fontSize: 16,
-    fontWeight: '600',
-    textAlign: 'center',
-  },
-  updateCurrentVersion: {
-    fontSize: 13,
-    marginTop: 2,
-  },
-  updateDivider: {
-    marginTop: SPACING.sm,
-  },
-  updateHintText: {
-    fontSize: 13,
-    lineHeight: 19,
-    marginTop: SPACING.md,
-    textAlign: 'center',
-  },
-  updateNewVersion: {
-    fontSize: 20,
-    fontWeight: '600',
-  },
-  updateVersionInfo: {
-    flex: 1,
-  },
-  updateVersionText: {
-    fontSize: 15,
-    lineHeight: 22,
-    textAlign: 'center',
   },
 });
